requirements.txt
playwright
pytest
pytest-bdd
pytest-html
pytest-xdist
allure-pytest